fix: RankTransferList.js 함수 부분 utils파일로 변경
